+++ assets/pages/Landing/Register.tsx
@@ -1,37 +1,9 @@
-import React, {useState} from 'react';
-import {ArrowLeft, Eye, EyeOff} from 'lucide-react';
-
-import {Button} from '@/components/ui/button';
-import {Input} from '@/components/ui/input';
-import {Label} from '@/components/ui/label';
-import {useSymfonyForm} from "@/hooks/useSymfonyForm";
+import React from 'react';
+import {ArrowLeft} from 'lucide-react';
 import {Link} from "react-router-dom";
 import {SymfonyFormWrapper} from "@/components/form/SymfonyForm";
 
 export default function Register() {
-
-    const {
-        errors,
-        handleSubmit,
-        handleChange,
-        elements
-    } = useSymfonyForm({
-        csrfNamespace: 'registration_form',
-        formKey: 'registration_form',
-        submitUrl: '/register',
-        initialData: {
-            email: '',
-            plainPassword: '',
-            _token: ''
-        }
-    });
-
-    const [showPassword, setShowPassword] = useState(false);
-
-
-    const togglePasswordVisibility = () => {
-        setShowPassword(!showPassword);
-    };
 
     return (
         <div className="m-auto w-full max-w-md space-y-8">
@@ -40,62 +12,7 @@ export default function Register() {
                 <p className="text-muted-foreground">Enter your details to sign up</p>
             </div>
             <div className="space-y-6">
-                <SymfonyFormWrapper formName='register' />
-
-                <hr />
-
-                <form className="space-y-4" onSubmit={handleSubmit}>
-                    <div className="space-y-2">
-                        <Label htmlFor="email">Email</Label>
-                        <Input
-                            id="email"
-                            type="email"
-                            name="email"
-                            placeholder="name@example.com"
-                            required
-                            onChange={handleChange}
-                            className={errors.email ? 'border-red-500' : ''}
-                        />
-                        {errors.email?.map((msg, i) => (
-                            <div key={i} className="text-red-500 text-sm">{msg}</div>
-                        ))}
-                    </div>
-                    <div className="space-y-2">
-                        <div className="flex items-center justify-between">
-                            <Label htmlFor="password">Password</Label>
-                        </div>
-                        <div className="relative">
-                            <Input
-                                id="password"
-                                name="plainPassword"
-                                type={showPassword ? 'text' : 'password'}
-                                placeholder="••••••••"
-                                required
-                                onChange={handleChange}
-                                className={errors.plainPassword ? 'border-red-500' : ''}
-                            />
-                            <Button
-                                type="button"
-                                variant="ghost"
-                                size="icon"
-                                className="absolute right-2 top-1/2 -translate-y-1/2"
-                                onClick={togglePasswordVisibility}
-                            >
-                                {showPassword ? <EyeOff className="h-4 w-4" /> : <Eye className="h-4 w-4" />}
-                                <span className="sr-only">Toggle password visibility</span>
-                            </Button>
-                        </div>
-                        {errors.plainPassword?.map((msg, i) => (
-                            <div key={i} className="text-red-500 text-sm">{msg}</div>
-                        ))}
-                    </div>
-
-                    {errors.registration_form?.map((msg, i) => (
-                        <div key={i} className="text-red-500 text-sm">{msg}</div>
-                    ))}
-
-                    <elements.SubmitButton text="Sign Up" submittingText="Submitting..."/>
-                </form>
+                <SymfonyFormWrapper formUrl='/register/render' onSuccess={() => alert('Success!')}/>
             </div>
             <div className="text-center text-sm">
                 Have an account already?{' '}
@@ -105,7 +22,7 @@ export default function Register() {
             </div>
             <div className="flex justify-center">
                 <a href="/" className="flex items-center text-sm text-muted-foreground hover:text-foreground">
-                    <ArrowLeft className="mr-2 h-4 w-4" />
+                    <ArrowLeft className="mr-2 h-4 w-4"/>
                     Back to home
                 </a>
             </div>
